[server] Add notice, process controller
